pagination
Added pagionation so 12 artworks are fetched at a time, fixing the wating time. Artwork history also works, so there will be 12 new, and the 12 last artworks viewed, but no more (optimization of cache). Main thing to fix; the user still has to wait between each fetch and has no visual indication that a fetch is happening.